@@ -1,23 +1,25 @@
 {
-  "name": "sag-ps-iot-pkg-layered-map-widget-plugin",
-  "version": "2.0.0",
+  "name": "layered-map-widget-plugin",
+  "version": "3.0.0",
   "description": "Show markers on a layer basis. Create layers based on filter criteria.",
   "scripts": {
     "start": "c8ycli server",
-    "start:shell": "c8ycli server -p 9326  -u https://entelios-dev.eu-latest.cumulocity.com/ --shell cockpit",
+    "start:shell": "c8ycli server -p 9326  -u https://your.tenant.cumulocity.com/ --shell cockpit",
     "build": "rimraf dist && c8ycli build",
-    "zip": "zip -r -q ./sag-ps-iot-pkg-layered-map-widget-plugin.zip ./dist/apps/sag-ps-iot-pkg-layered-map-widget-plugin/*",
     "lint": "eslint . --ext js,ts --quiet",
     "format": "eslint . --ext js,ts --quiet --fix",
-    "build:ci": "npm run build -- --ci",
-    "deploy": "c8ycli deploy",
+    "build:ci": "npm run build --ci",
+    "deploy": "c8ycli deploy -u https://your-tenant.cumulocity.com/",
     "postinstall": "ngcc"
   },
   "keywords": [
     "Cumulocity",
     "Plugin",
     "map",
-    "widget"
+    "widget",
+    "leaflet",
+    "layered",
+    "WMS"
   ],
   "author": "Hendrik Naether",
   "license": "Apache 2.0",
@@ -32,17 +34,18 @@
     "@angular/platform-browser-dynamic": "14.0.6",
     "@angular/router": "14.0.6",
     "@angular/upgrade": "14.0.6",
-    "@types/lodash": "^4.14.194",
     "@asymmetrik/ngx-leaflet": "^14.0.1",
-    "@c8y/client": "1017.0.267",
-    "@c8y/ngx-components": "1017.0.267",
-    "@c8y/style": "1017.0.267",
+    "@c8y/client": "1018.0.233",
+    "@c8y/ngx-components": "1018.0.233",
+    "@c8y/style": "1018.0.233",
     "@ngx-translate/core": "14.0.0",
-    "@types/leaflet": "^1.9.0",
-    "angular-resize-event": "3.2.0",
+    "@types/leaflet": "^1.9.8",
+    "@types/lodash": "^4.14.194",
     "jsoneditor": "^9.10.2",
-    "leaflet": "^1.9.3",
+    "leaflet": "^1.9.4",
+    "leaflet-wms-header": "^1.0.13",
     "moment": "^2.29.4",
+    "ngx-bootstrap": "^9.0.0",
     "rxjs": "~6.6.3",
     "zone.js": "~0.11.7"
   },
@@ -52,18 +55,17 @@
     "@angular/language-service": "14.0.6",
     "@angular/localize": "14.0.6",
     "@angular/service-worker": "14.0.6",
-    "@c8y/cli": "1017.0.262",
-    "@types/jest": "^27.0.3",
+    "@c8y/cli": "1018.0.233",
+    "@types/jest": "^28.1.8",
     "@typescript-eslint/eslint-plugin": "^5.42.1",
     "@typescript-eslint/parser": "^5.42.1",
-    "eslint": "^8.7.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
+    "eslint": "^8.18.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "html-loader": "4.1.0",
-    "jest": "^27.4.5",
-    "jest-preset-angular": "^11.0.1",
-    "prettier": "^2.5.1",
-    "style-loader": "3.3.1",
+    "jest": "^28.1.3",
+    "jest-preset-angular": "^12.2.0",
+    "prettier": "^2.7.1",
     "typescript": "4.7.4"
   },
   "c8y": {
@@ -76,7 +78,7 @@
       "tabsHorizontal": true,
       "isPackage": true,
       "noAppSwitcher": true,
-      "requiredPlatformVersion": "1017.0",
+      "requiredPlatformVersion": "1018.0",
       "package": "plugin",
       "exports": [
         {
