@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { ModalLabels } from '@c8y/ngx-components';
+import { ModalLabels, ModalService, Status } from '@c8y/ngx-components';
 import { IconSelectorService } from '@c8y/ngx-components/icon-selector';
 import { BsModalRef } from 'ngx-bootstrap/modal';
 import { Subject } from 'rxjs';
@@ -8,22 +8,39 @@ import {
   DeviceFragmentLayerConfig,
   isDeviceFragmentLayerConfig,
   isQueryLayerConfig,
+  isWebMapServiceLayerConfig,
   QueryLayerConfig,
+  WebMapServiceLayerConfig,
 } from '../layered-map-widget.model';
+import { TenantOptionCredentialsService } from '../../layered-map-widget-plugin/service/tenant-option-credentials.service';
 
 @Component({ templateUrl: './layer-modal.component.html' })
 export class LayerModalComponent {
   title = 'Create layer';
-  closeSubject: Subject<BasicLayerConfig> = new Subject();
+  closeSubject: Subject<DeviceFragmentLayerConfig | QueryLayerConfig | WebMapServiceLayerConfig> =
+    new Subject();
   labels: ModalLabels = { ok: 'Create', cancel: 'Cancel' };
-  layer: BasicLayerConfig = { name: '', color: '', icon: '' };
+  layer: Partial<DeviceFragmentLayerConfig | QueryLayerConfig | WebMapServiceLayerConfig> = {
+    name: '',
+    color: '',
+    icon: '',
+    enablePolling: 'true',
+    pollingInterval: 60,
+  };
 
-  type: 'DeviceFragmentLayerConfig' | 'QueryLayerConfig' | 'Unset' = 'Unset';
+  type: 'DeviceFragmentLayerConfig' | 'QueryLayerConfig' | 'Unset' | 'WebMapServiceLayer' = 'Unset';
   queryType: 'Alarm' | 'Inventory' | 'Event';
 
-  constructor(public bsModalRef: BsModalRef, private iconSelector: IconSelectorService) {}
+  protected wmsCredentials = { username: '', password: '' };
 
-  setLayer(layer: DeviceFragmentLayerConfig | QueryLayerConfig) {
+  constructor(
+    public bsModalRef: BsModalRef,
+    private iconSelector: IconSelectorService,
+    private modal: ModalService,
+    private tenantOptionCredentials: TenantOptionCredentialsService
+  ) {}
+
+  setLayer(layer: BasicLayerConfig) {
     this.layer = layer;
     this.title = 'Edit layer';
     this.labels = { ok: 'Update', cancel: 'Cancel' };
@@ -32,23 +49,32 @@ export class LayerModalComponent {
     } else if (isQueryLayerConfig(layer)) {
       this.type = 'QueryLayerConfig';
       this.queryType = layer.type;
+    } else if (isWebMapServiceLayerConfig(layer)) {
+      this.type = 'WebMapServiceLayer';
+      if (layer.token) {
+        this.tenantOptionCredentials.getCredentials(layer.token).then((creds) => {
+          this.wmsCredentials = creds;
+        });
+      }
     }
   }
 
   async openIconModal() {
-    const icon = await this.iconSelector.selectIcon();
+    const icon = await this.iconSelector.selectIcon({ currentSelection: this.layer.icon });
     if (icon) {
       this.layer.icon = icon;
     }
   }
 
   changeType(type: string) {
-    const { name, color, icon } = this.layer;
+    const { name, color, icon, pollingInterval, enablePolling } = this.layer;
     if (type === 'DeviceFragmentLayerConfig') {
       this.layer = {
         name,
         color,
         icon,
+        pollingInterval,
+        enablePolling,
         ...{ fragment: '', value: '', device: { id: '', name: '' } },
       } as DeviceFragmentLayerConfig;
       this.type = type;
@@ -57,6 +83,8 @@ export class LayerModalComponent {
         name,
         color,
         icon,
+        pollingInterval,
+        enablePolling,
         ...{ type: 'Alarm', filter: {} },
       } as QueryLayerConfig;
       this.type = 'QueryLayerConfig';
@@ -66,6 +94,8 @@ export class LayerModalComponent {
         name,
         color,
         icon,
+        pollingInterval,
+        enablePolling,
         ...{ type: 'Event', filter: {} },
       } as QueryLayerConfig;
       this.type = 'QueryLayerConfig';
@@ -75,11 +105,60 @@ export class LayerModalComponent {
         name,
         color,
         icon,
+        pollingInterval,
+        enablePolling,
         ...{ type: 'Inventory', filter: {} },
       } as QueryLayerConfig;
       this.type = 'QueryLayerConfig';
       this.queryType = 'Inventory';
+    } else if (type === 'WebMapServiceLayer') {
+      this.layer = {
+        name,
+        color,
+        icon,
+        pollingInterval,
+        enablePolling,
+        ...{ type: 'ExternalGIS', url: '', wmsLayers: [{ name: '' }] },
+      } as WebMapServiceLayerConfig;
+      this.type = 'WebMapServiceLayer';
     }
+  }
+
+  onWmsURLChange(url?: string) {
+    if (url?.length > 0) {
+      const params = new URLSearchParams(decodeURI(url));
+      if (params.has('layers')) {
+        const layers = params.get('layers')?.split(',');
+        const existingLayers = (<WebMapServiceLayerConfig>this.layer).wmsLayers;
+        for (const layer of layers) {
+          if (!existingLayers.find((l) => l.name === layer)) {
+            existingLayers.push({ name: layer });
+          }
+        }
+      }
+    }
+  }
+
+  addWMSLayer() {
+    (<WebMapServiceLayerConfig>this.layer).wmsLayers.push({ name: '' });
+  }
+
+  removeWMSLayer(index: number) {
+    (<WebMapServiceLayerConfig>this.layer).wmsLayers.splice(index, 1);
+  }
+
+  clearAllPasswords() {
+    this.modal
+      .confirm('Clear passwords', 'Are you sure you want to clear all passwords?', Status.DANGER)
+      .then((result) => {
+        if (result) this.onClearPasswordsConfirmation();
+      });
+  }
+
+  private onClearPasswordsConfirmation() {
+    this.tenantOptionCredentials.clearAllCredentials();
+    this.wmsCredentials = { username: '', password: '' };
+    delete (<WebMapServiceLayerConfig>this.layer).token;
   }
 
   // - MODAL section
@@ -91,6 +170,21 @@ export class LayerModalComponent {
 
   // called if save is pressed
   onClose(): void {
-    this.closeSubject.next(this.layer);
+    if (
+      this.type === 'WebMapServiceLayer' &&
+      this.wmsCredentials.username.length &&
+      this.wmsCredentials.password.length
+    ) {
+      const creds = {
+        username: this.wmsCredentials.username,
+        password: this.wmsCredentials.password,
+      };
+      this.tenantOptionCredentials.saveCredentials(creds).then((token) => {
+        (<WebMapServiceLayerConfig>this.layer).token = token;
+        this.closeSubject.next(this.layer as any);
+      });
+    } else {
+      this.closeSubject.next(this.layer as any);
+    }
   }
 }
