@@ -1,11 +1,13 @@
 import { Injectable } from '@angular/core';
 import { IManagedObject } from '@c8y/client';
-import { LatLng, latLng, LayerGroup, Marker } from 'leaflet';
-import { get, has, isEmpty, set } from 'lodash';
+import { FeatureGroup, LatLng, latLng, Marker } from 'leaflet';
+import { flatten, get, has, isEmpty, set } from 'lodash';
 import {
+  BasicLayerConfig,
   DeviceFragmentLayerConfig,
   isDeviceFragmentLayerConfig,
   isQueryLayerConfig,
+  isWebMapServiceLayerConfig,
   LayerConfig,
   MyLayer,
   PollingDelta,
@@ -24,35 +26,41 @@ export class LayerService {
     private selectedDevicesService: SelectedDevicesService
   ) {}
 
-  createLayers(configs: LayerConfig[]) {
+  createLayers(configs: LayerConfig<BasicLayerConfig>[]) {
     return Promise.all(configs.map((cfg) => this.createLayer(cfg)));
   }
 
-  async createLayer(setup: LayerConfig) {
+  async createLayer(setup: LayerConfig<BasicLayerConfig>) {
     const layer = Object.assign(new MyLayer(), setup);
 
     if (isQueryLayerConfig(setup.config)) {
       const config = setup.config;
       if (config.type === 'Alarm') {
-        this.queryLayerService.fetchByAlarmQuery(config.filter).then((devices) => {
-          layer.devices = devices.map((d) => d.id);
-          devices.forEach((d) => {
-            this.updatePosition(layer, d.id, d.c8y_Position);
-            this.updateMarkerIcon(d.id, layer, d.c8y_ActiveAlarmsStatus);
+        layer.initialLoad = this.queryLayerService
+          .fetchByAlarmQuery(config.filter)
+          .then((devices) => {
+            layer.devices = devices.map((d) => d.id);
+            devices.forEach((d) => {
+              this.updatePosition(layer, d.id, d.c8y_Position);
+              this.updateMarkerIcon(d.id, layer, d.c8y_ActiveAlarmsStatus);
+            });
           });
-        });
       } else if (config.type === 'Inventory') {
-        this.queryLayerService.fetchByInventoryQuery(config.filter).then((devices) => {
-          layer.devices = devices.map((d) => d.id);
-          devices.forEach((d) => this.updatePosition(layer, d.id, d.c8y_Position));
-        });
+        layer.initialLoad = this.queryLayerService
+          .fetchByInventoryQuery(config.filter)
+          .then((devices) => {
+            layer.devices = devices.map((d) => d.id);
+            devices.forEach((d) => this.updatePosition(layer, d.id, d.c8y_Position));
+          });
       } else if (config.type === 'Event') {
-        this.queryLayerService.fetchByEventQuery(config.filter).then((devices) => {
-          layer.devices = devices.map((d) => d.id);
-          devices.forEach((d) => this.updatePosition(layer, d.id, d.c8y_Position));
-        });
+        layer.initialLoad = this.queryLayerService
+          .fetchByEventQuery(config.filter)
+          .then((devices) => {
+            layer.devices = devices.map((d) => d.id);
+            devices.forEach((d) => this.updatePosition(layer, d.id, d.c8y_Position));
+          });
       }
-    } else {
+    } else if (isDeviceFragmentLayerConfig(setup.config)) {
       const config = setup.config as DeviceFragmentLayerConfig;
       const devices = await this.selectedDevicesService.getDevices(config.device);
       const matches = this.getMatches(setup.config, devices || []);
@@ -64,13 +72,18 @@ export class LayerService {
         .forEach((d) => layer.coordinates.set(d.id, latLng(d.c8y_Position)));
 
       this.createLayerGroup(layer);
+      layer.initialLoad = Promise.resolve();
+    } else if (isWebMapServiceLayerConfig(setup.config)) {
+      layer.config.enablePolling = 'false';
+      layer.config.icon = 'globe1';
+      layer.initialLoad = Promise.resolve();
     }
 
     return layer;
   }
   updateMarkerIcon(
     deviceId: string,
-    layer: MyLayer & LayerConfig,
+    layer: MyLayer & LayerConfig<BasicLayerConfig>,
     status: {
       critical?: number;
       major?: number;
@@ -101,7 +114,9 @@ export class LayerService {
         this.updateMarkerIcon(mo.id, layer, mo.c8y_ActiveAlarmsStatus);
       }
       const marker = this.updatePosition(layer, mo.id, mo.c8y_Position);
-      this.popupService.getPopupComponent(marker).onUpdate(mo);
+      if (marker) {
+        this.popupService.getPopupComponent(marker).onUpdate(mo);
+      }
     }
   }
 
@@ -137,12 +152,12 @@ export class LayerService {
       return marker;
     });
 
-    layer.group = new LayerGroup(markers);
+    layer.group = new FeatureGroup(markers);
   }
 
   private createMarker(deviceId: string, coordinate: LatLng, layer: MyLayer) {
-    // TODO: add layer specific stuff here
-    const icon = this.markerIconService.getIcon(layer.config.icon);
+    const color = layer.config.color?.length ? layer.config.color : '#fff';
+    const icon = this.markerIconService.getIcon(layer.config.icon, '', color);
     const popup = this.popupService.getPopup({ deviceId, layer });
 
     const marker = new Marker(coordinate, {
@@ -161,8 +176,15 @@ export class LayerService {
     return devices;
   }
 
-  private updatePosition(layer: MyLayer, id: string, position: any) {
+  private updatePosition(layer: MyLayer, id: string, position: any): Marker | null {
     let marker: Marker<any> = null;
+    if (!position) {
+      if (layer.coordinates.has(id)) {
+        layer.coordinates.delete(id);
+        layer.markerCache.delete(id);
+      }
+      return marker;
+    }
     // we haven't had any position yet
     if (!layer.coordinates.has(id)) {
       const coordinate = latLng(position);
@@ -180,5 +202,10 @@ export class LayerService {
       }
     }
     return marker;
+  }
+
+  extractMinMaxBounds(allLayers: MyLayer[]) {
+    const markers = flatten(allLayers.map((l) => [...l.markerCache.values()]));
+    return new FeatureGroup(markers).getBounds();
   }
 }
